FEATURE: Bulk member management

+++ app/assets/javascripts/discourse/app/controllers/group-index.js
@@ -54,6 +54,7 @@ export default Controller.extend({
         "application.showFooter":
           this.model.members.length >= this.model.user_count,
         loading: false,
+        bulkSelection: [],
       });
     });
   },
@@ -114,38 +115,55 @@ export default Controller.extend({
 
   @action
   actOnSelection(selection, actionId) {
-    const userIds = (selection || []).map((member) => member.id);
+    if (!selection || selection.length === 0) {
+      return;
+    }
 
     switch (actionId) {
       case "removeMembers":
         return ajax(`/groups/${this.model.id}/members.json`, {
           type: "DELETE",
-          data: { user_ids: userIds },
-        }).then(() => this.model.findMembers(this.memberParams, true));
+          data: { user_ids: selection.map((u) => u.id).join(",") },
+        }).then(() => {
+          this.model.findMembers(this.memberParams, true);
+          this.set("isBulk", false);
+        });
 
       case "makeOwners":
         return ajax(`/admin/groups/${this.model.id}/owners.json`, {
           type: "PUT",
-          data: { user_ids: userIds },
-        }).then(() => selection.forEach((s) => s.set("owner", true)));
+          data: {
+            group: { usernames: selection.map((u) => u.username).join(",") },
+          },
+        }).then(() => {
+          selection.forEach((s) => s.set("owner", true));
+          this.set("isBulk", false);
+        });
 
       case "removeOwners":
         return ajax(`/admin/groups/${this.model.id}/owners.json`, {
           type: "DELETE",
-          data: { user_ids: userIds },
-        }).then(() => selection.forEach((s) => s.set("owner", false)));
+          data: {
+            group: { usernames: selection.map((u) => u.username).join(",") },
+          },
+        }).then(() => {
+          selection.forEach((s) => s.set("owner", false));
+          this.set("isBulk", false);
+        });
 
-      case "makeAllPrimary":
+      case "setPrimary":
+      case "unsetPrimary":
+        const primary = actionId === "setPrimary";
         return ajax(`/admin/groups/${this.model.id}/primary.json`, {
           type: "PUT",
-          data: { user_ids: userIds },
-        }).then(() => selection.forEach((s) => s.set("primary", true)));
-
-      case "removeAllPrimary":
-        return ajax(`/admin/groups/${this.model.id}/primary.json`, {
-          type: "DELETE",
-          data: { user_ids: userIds },
-        }).then(() => selection.forEach((s) => s.set("primary", false)));
+          data: {
+            group: { usernames: selection.map((u) => u.username).join(",") },
+            primary,
+          },
+        }).then(() => {
+          selection.forEach((s) => s.set("primary", primary));
+          this.set("isBulk", false);
+        });
     }
   },
 
@@ -178,7 +196,7 @@ export default Controller.extend({
   toggleBulkSelect() {
     this.setProperties({
       isBulk: !this.isBulk,
-      bulkSelection: null,
+      bulkSelection: [],
     });
   },
